fix(nn): round TPA's fused product before the add — FMA contraction broke the parity pin
TestTPAForwardFusedBitExactVsDispatch fails on main. It pins TPA's fused
inference contract byte-for-byte against the einsum dispatch path, and the two
differ by exactly one ulp: fused 0.14259979878693926, dispatch ...24.

CAUSE. The fused loop is written `s += a[i] * b[j]`, which the Go compiler
CONTRACTS into a fused multiply-add on arm64 — 32 FMADDD in that one loop —
and does not contract on amd64. An FMA rounds once where a separate multiply
and add round twice, so the fused path is a different (more accurate) number on
one architecture only. Writing the product through an explicit conversion forces
the intermediate rounding the einsum engine performs, and the pin holds again.

This is why the PR that added the test merged green: its CI runs on amd64, where
no contraction happens and both paths agree. The failure is invisible there and
unavoidable on any arm64 machine. The repo already had one recorded instance of
this exact mechanism — rl/softupdate_parity_test.go documents a one-ulp failure
from FMA contraction differing between expression shapes — so the failure mode
was known, just not guarded.

NOTE FOR THE REST: an intermediate VARIABLE does not help. Assigning the product
to a local left all 32 FMADDD in place; only the explicit float64(...) conversion
suppresses contraction, which is what the Go spec actually guarantees.

SCOPE. This fixes TPA only, and main still has nine failures in the same family
(KAN, MTAHeadConv, MemForward, plus six that predate the recent merge batch:
DeltaNet, EMAUpdate, GLA, HGRNSeq, RGLRUSeq, TitansLinear). Their fused loops do
not have the same shape and the same edit does not fix them, so each needs its
own diagnosis rather than a blanket rewrite.

+++ nn/tpa.go
@@ -230,7 +230,7 @@ func (m *TPA) contractFusedF64(a, b *tensor.Tensor, r, t int) (*tensor.Tensor, e
 			for d := 0; d < dh; d++ {
 				var s float64
 				for rr := 0; rr < r; rr++ {
-					s += as[aBase+rr*heads+h] * bs[bBase+rr*dh+d]
+					s += float64(as[aBase+rr*heads+h] * bs[bBase+rr*dh+d])
 				}
 				os[oBase+h*dh+d] = s * inv
 			}
